refactor: mapper object 제거 및 적용
