+++ app/src/main/java/com/teamwss/websoso/domain/usecase/GetFeedsUseCase.kt
@@ -1,7 +1,7 @@
 package com.teamwss.websoso.domain.usecase
 
 import com.teamwss.websoso.data.repository.DefaultFeedRepository
-import com.teamwss.websoso.domain.mapper.FeedMapper.toDomain
+import com.teamwss.websoso.domain.mapper.toDomain
 import com.teamwss.websoso.domain.model.Feeds
 import javax.inject.Inject
